fix(jest): couldn't load config cause it depends on Utils, and utils depends on config in some way
